+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Improvements
 - Added support for ATT&CK Workbench as a datasource
 - Added parsing for CAPEC IDs in AttackToExcel
+- Added support for data sources and data components when generating layers
 
 # v1.4.0 - 10/21/2021
 ## Fixes